requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
distro==1.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
jiter==0.5.0
openai==1.43.0
pydantic==2.8.2
pydantic_core==2.20.1
pyTelegramBotAPI==4.22.1
python-dotenv==1.0.1
python-telegram-bot==21.5
requests==2.32.3
sniffio==1.3.1
telegram==0.0.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
